feat: refactor notebook components and introduce LeftBar for improved layout
