+++ DashAI/front/src/components/notebooks/RightBar.jsx
@@ -1,18 +1,5 @@
-import { Box } from "@mui/material";
+import SideBar from "../threeSectionLayout/SideBar";
 
-export default function RightBar({ children }) {
-  return (
-    <Box
-      display={"flex"}
-      height={"100%"}
-      width={"100%"}
-      flexDirection={"column"}
-      justifyContent={"flex-start"}
-      overflow={"auto"}
-      bgcolor={"background.box"}
-      borderRadius={2}
-    >
-      {children}
-    </Box>
-  );
+export default function RightBar() {
+  return <SideBar></SideBar>;
 }
