+++ DashAI/front/src/components/threeSectionLayout/CenterBox.jsx
