+++ DashAI/front/src/components/threeSectionLayout/Footer.jsx
